[Client] Use uint16_t in Packager

+++ client/includes/Packager.hh
@@ -17,9 +17,9 @@ public:
 	Packager();
 	virtual ~Packager();
 
-	static void createMovementPackage(uint8_t _playerID, uint8_t _x,	uint8_t _y);
-	static void createShotPackage(uint8_t _playerID, uint8_t _type, uint8_t _x, uint8_t _y);
-	
+	static void createMovementPackage(uint8_t _playerID, uint16_t _x,	uint16_t _y);
+	static void createShotPackage(uint8_t _playerID, uint8_t _type, uint16_t _x, uint16_t _y);
+
 	static void createGameListPackage();
 	static void createPlayerListPackage();
